Split behaviors and phenomena
