+++ lib/doro/phenomena/phenomena.ex
@@ -1,0 +1,30 @@
+defmodule Doro.Phenomena do
+  @moduledoc """
+  Functions for running autonomous behaviors that effect changes in the
+  game state without any player interaction
+  """
+
+  require Logger
+  use GenServer
+
+  def start_link do
+    GenServer.start_link(__MODULE__, [], name: :phenomena)
+  end
+
+  def init(args) do
+    Phoenix.PubSub.subscribe(Doro.PubSub, "heartbeat")
+    {:ok, args}
+  end
+
+  def handle_info({:heartbeat, t}, state) do
+    [
+      Doro.Phenomena.Clock,
+      Doro.Phenomena.VascularPlant
+    ]
+    |> Enum.each(fn phenom ->
+      Task.start(fn -> phenom.tick(t) end)
+    end)
+
+    {:noreply, state}
+  end
+end
