+++ lib/doro/phenomena/vascular_plant.ex
@@ -1,0 +1,4 @@
+defmodule Doro.Phenomena.VascularPlant do
+  def tick(t) do
+  end
+end
